Merge branch 'tomato-ND-KUP24' into tomato-KUP24-usbmod-base

+++ release/src/linux/linux/include/asm-ppc/posix_types.h
@@ -1,6 +1,3 @@
-/*
- * BK Id: SCCS/s.posix_types.h 1.5 05/17/01 18:14:25 cort
- */
 #ifndef _PPC_POSIX_TYPES_H
 #define _PPC_POSIX_TYPES_H
 
@@ -11,7 +8,7 @@
  */
 
 typedef unsigned int	__kernel_dev_t;
-typedef unsigned int	__kernel_ino_t;
+typedef unsigned long	__kernel_ino_t;
 typedef unsigned int	__kernel_mode_t;
 typedef unsigned short	__kernel_nlink_t;
 typedef long		__kernel_off_t;
@@ -75,7 +72,7 @@ static __inline__ void __FD_CLR(unsigned long fd, __kernel_fd_set *fdsetp)
 
 #undef __FD_ISSET
 static __inline__ int __FD_ISSET(unsigned long fd, __kernel_fd_set *p)
-{ 
+{
 	unsigned long _tmp = fd / __NFDBITS;
 	unsigned long _rem = fd % __NFDBITS;
 	return (p->fds_bits[_tmp] & (1UL<<_rem)) != 0;
